Added event handler to planetoxygenservice and connected event 'oxygenUpdate' to OxygenDisplay to trigger update.

+++ source/core/src/main/com/csse3200/game/services/OxygenLevel.java
@@ -26,5 +26,5 @@ public interface OxygenLevel {
      * Gets the oxygen present as a percentage of the upper limit (Float 0.0 - 100.0).
      * @return The oxygen present as a percentage of the upper limit.
      */
-    float getOxygenPercentage();
+    int getOxygenPercentage();
 }
